DEVELOPMENT_2026_08_13_1924, <COMPILE_SUCCESS> <RUNTIME_TEST>. PHASE: testing
implemented:
 - further testing and changes of 'unsigned long long' as "priaseId" parameter.

cmake -B build
-- The CXX compiler identification is GNU 15.2.0
-- Detecting CXX compiler ABI info
-- Detecting CXX compiler ABI info - done
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Detecting CXX compile features - done
-- Configuring done (0.4s)
-- Generating done (0.0s)
-- Build files have been written to: /home/brenton-james/CLionProjects/OpenEpiCentre_BackEndUbuntu

cmake --build build
[  2%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/CLIB_OpenEpiCentre.cpp.o
[  4%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Framework.cpp.o
[  7%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Global.cpp.o
[  9%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/CLIB_OpenEpiCentre_STRUCT__Input.cpp.o
[ 11%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/CLIB_OpenEpiCentre_STRUCT__Output.cpp.o
[ 14%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/CLIB_OpenEpiCentre_STRUCT__Concurrent.cpp.o
[ 16%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/CLIB_OpenEpiCentre_STRUCT__User_Input.cpp.o
[ 19%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/CLIB_OpenEpiCentre_STRUCT__User_Output.cpp.o
[ 21%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/CLIB_OpenEpiCentre_STRUCT__User_Algorithm.cpp.o
[ 23%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Algorithm_praise0.cpp.o
[ 26%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Input_praise0.cpp.o
[ 28%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Output_praise0.cpp.o
[ 30%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Algorithm_praise1.cpp.o
[ 33%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Input_praise1.cpp.o
[ 35%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Output_praise1.cpp.o
[ 38%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Algorithm_praise2.cpp.o
[ 40%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Input_praise2.cpp.o
[ 42%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Output_praise2.cpp.o
[ 45%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Algorithm_praise3.cpp.o
[ 47%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Input_praise3.cpp.o
[ 50%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/structs/praise_sets/CLIB_OpenEpiCentre_STRUCT__Output_praise3.cpp.o
[ 52%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__App.cpp.o
[ 54%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Algorithms.cpp.o
[ 57%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Data.cpp.o
[ 59%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Data__Control.cpp.o
[ 61%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Execute.cpp.o
[ 64%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerInputReceive/src/CLIB_WriteEnableForThreadsAt_ServerInputReceive.cpp.o
[ 66%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerInputReceive/src/WriteEnableForThreadsAt_ServerInputReceive__Framework.cpp.o
[ 69%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerInputReceive/src/WriteEnableForThreadsAt_ServerInputReceive__Global.cpp.o
[ 71%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerInputReceive/src/WriteEnableForThreadsAt_ServerInputReceive__App.cpp.o
[ 73%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerInputReceive/src/WriteEnableForThreadsAt_ServerInputReceive__Control.cpp.o
[ 76%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_LaunchQueForThreadsAt_Server/src/LaunchEnableForConcurrentThreadsAt_Server__Framework.cpp.o
[ 78%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_LaunchQueForThreadsAt_Server/src/LaunchEnableForConcurrentThreadsAt_Server__Global.cpp.o
[ 80%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_LaunchQueForThreadsAt_Server/src/LaunchEnableForConcurrentThreadsAt_Server__App.cpp.o
[ 83%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_LaunchQueForThreadsAt_Server/src/LaunchEnableForConcurrentThreadsAt_Server__Control.cpp.o
[ 85%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_LaunchQueForThreadsAt_Server/src/CLIB_LaunchEnableForConcurrentThreadsAt_Server.cpp.o
[ 88%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerOutputSend/src/CLIB_WriteEnableForThreadsAt_ServerOutputSend.cpp.o
[ 90%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerOutputSend/src/WriteEnableForThreadsAt_ServerOutputSend__Framework.cpp.o
[ 92%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerOutputSend/src/WriteEnableForThreadsAt_ServerOutputSend__Global.cpp.o
[ 95%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/libs/CLIB_WriteQueForThreadsAt_ServerOutputSend/src/WriteEnableForThreadsAt_ServerOutputSend__Control.cpp.o
[ 97%] Building CXX object CMakeFiles/CLIB_OpenEpiCentre.dir/src/engine/CLIB_OpenEpiCentre__Execute__Control.cpp.o
[100%] Linking CXX static library libCLIB_OpenEpiCentre.a
[100%] Built target CLIB_OpenEpiCentre

+++ CLIB_OpenEpiCentre/src/CLIB_OpenEpiCentre.cpp
@@ -291,7 +291,7 @@
 	}
 	unsigned char* CLIBOpenEpiCentre::CLIBConcurrentServerIO::CLIBConcurrentServerIO__io_PRAISE_get_MetaData_PraiseEventId()
 	{
-		unsigned long* temp;
+		unsigned long long* temp;
 		*temp = UINT64_MAX;
 		if (CLIBConcurrentServerIO__stat_REG_get_ptr_CLIB_OpenEpiCentre_isFlagINSTANTIATED()[12] == false) {
 			*temp = CLIBConcurrentServerIO__stat_CLASS_get_ptr_CLIB_OpenEpiCentre_Framework()->dyn_CLASS_get_ptr_CLIB_OpenEpiCentre_App()->dyn_CLASS_get_ptr_Data()->dyn_REG_get_ptr_Item_array_Of_doubleBuffer_Output_WRITE(CLIBConcurrentServerIO__stat_CLASS_get_ptr_CLIB_OpenEpiCentre_Framework())->dyn_REG_get_ptr_CLIB_OpenEpiCentre_STRUCT__Output_praiseEventId();
